Send application context for each application line.
This enables support for GitHub links in the Honeybadger UI. The otp
application can be configured like:

Application.put_env(:honeybadger, :app, :crywolf)

The caveat is that files referencing dependencies of the application
(such as phoenix, postgrex, etc.) are also prefixed even though they do
not reside under source control. That means that lines such as
"lib/phoenix/router.ex:2:in `call'" in the backtrace
will be linked to GitHub files which don't exist. Ideally we'll find a
better way to detect true project files within the application so that
we can display GitHub links only for files which actually reside under
source control.