Add Logger if use_logger config is true
Users can pass a `use_logger` settings to the config.
This setting defaults to false. If it true we add the
Honeybadger.Logger handler to Erlang's error_logger.

Note we are not using Elixir's Logger. This is due to the fact
that we can get more info from Erlang's error_logger right now.