Only handle error messages that are lists
Some supervisor messages were making it through and the message types
were of tuple. This caused the code that looks into the message using
Dict to fail.

This change uses a guard to check that the message is a
list and matches specifically on an error type of `:error_report`. I've
also added a catch all message type after the main handle_event to
swallow all other messages.