Check for and prefer MIX_ENV variable before raising exception
Since all deps for a project get compiled with their `env` set to `prod`
we need to inform our users to either set `environment_name` in their
config files for Honeybadger. If the `environment_name` is not set we
will raise an exception.

This change also checks for the `MIX_ENV` environment variable before
raising an exception and gives precedence to the `MIX_ENV` environment
variable over the value specific in config.exs files.