Use Mix.env directly and always compare as atom
Now that the `environment_name` is defined within `mix.exs` by default
we can be sure that `Mix.env/0` is available. Using `Mix.env/0` directly
means that we will have an atom based env most of the time. In rare
cases where we can't be sure (i.e., in a release using `system` tuples)
we convert a binary env to an atom.

Closes #94