Generate smarter stacktraces from custom notify
When `Honeybadger.notify/1|2` is called from within a `try` expression
the exception stacktrace as provided by `System.stacktrace/0` is
available. When `System.stacktrace/0` is called _outside_ of a `try`
expression it simply returns `[]`, and we can fall back to using
`Process.info/2` to generate a messier stacktrace.

Because the `notify` function is typically used within a `try`
expression this provides much better default stacktraces.

Closes #99