Add explicit support for sending string noticies
While it was possible to send an arbitrary notice by constructing an
error manually, it wasn't possible to pass a string to the `notify`
function. Sending a notice directly from a string is more convenient,
and it make the exact arguments that the `notify` function accepts more
obvious.

This refactors the `Honeybadger.Notice` module a little to guard the
accepted types a bit more. Part of the guarding process involved adding
typespecs for the components that make up a `Honeybadger.Notice` struct.

Addresses #92