Respect URL filtering in cgi_data (#105)
* Reorganize structure of filter modules

With all of the modules in the same file it masks how many modules there
are and forces explicit ordering within the file. This extracts each of
the modules into its own file and uses namespacing to indicate the
dependency structure.

* Use with_config/2 helper for app env testing

* Filter url data from cgi_data when enabled

The notice's `cgi_data` map contains several fields that expose URL
data. That data is no longer retained when `filter_url_data` is enabled.

Closes #102

* Note filtering changes and fixes in the CHANGELOG