Add Honeybadger.clear_context
I am trying to use Honeybadger.context/1 inside a Honeydew worker, which
is a long-lived process that processes multiple jobs. Because these
workers are long-lived, when a worker starts a new job the context from
the previous job is carried over until the new job overwrites it.

This adds a `Honeybadger.clear_context/0` function which could be called
any time you'd like the context to be completely reset for the current
process. I can't think of any additional issues this would cause.

I also added docs and typespecs for `context/0` and `context/1`.