Fetch peer information for multiple plug versions (#157)
Plug 1.6 stopped exposing the `Conn` struct's `peer` field. This may be
fixed in a newer version, but it is deprecated anyhow, so we're future
proofing things here.

The request info is now built using `Conn.get_peer_data/1` when
available, falling back to direct field access otherwise.

Additional discussion in #155.