Remove use of System.stracktrace/0 outside of try
This adds tests for Erlang 21.0 and Elixir 1.7.0 to the build matrix.
One of the big changes is the introduction of `__STACKTRACE__` and the
deprecation of `System.stacktrace/0`.

To prevent compile time warnings moving forward we are removing the
fallback `System.stacktrace/0` call inside of `honeybadger.ex`. The call
was made outside of a `try/catch` and wasn't guaranteed to return the
correct stacktrace anyhow.