Guard against nil returns from Process.info/2
When a process is no longer alive any calls to `Process.info/2` will
return `nil` instead of the expected tuple. Inside of the fallback
clause for `backtrace/1` we always matched on a stacktrace tuple and
didn't account for the possibility of a `nil` return. This change makes
the backtrace generation safer by using a `case` instead of a fixed
match.

During my experimentation I wasn't able to call `backtrace/1` from a
process and have the process die before `Process.info/2` was called. The
case clause is defensive and should cover the known failure case, but
it seems to be untestable.

Closes #172