Migrate to Elixir 1.7+ Logger (#204)
* Migrate error reporting to use new logger

In Elixir 1.7+ and Erlang 21+ we can use the standardized error logger
to consolidate and simplify error reporting. The biggest resulting
change is that `context` _must_ be stored as Logger `metadata`.

* Require explicitly passing stacktrace to notify

The use of `System.stacktrace/0` outside of rescue/catch is deprecated
and automatically extracting the stacktrace is unreliable. To encourage
reporting within a rescue/catch block `Honeybadger.notify/3` now accepts
`stacktrace` as the second argument and *will not* automatically extract
the current processes' stacktrace.

* Extract additional context from logger messages

* Update version, changelog and build matrix