Normalize string errors before exception blame (#226)
We need to convert string based errors to an exception before calling
`Exception.blame/3` on exceptions. Calling `blame` requires a proper
stacktrace to work. However, normalization took place inside the
`Notice` module which was only passed the Honeybadger specific
backtrace.

In order to centralize normalization and improve the blame behavior all
of the conversion now happens in `Notice.new/3`. As part of this effort
the typespecs, guards and function signatures have been improved.

The `Backtrace` and `Notice` modules are also no longer public. There
wasn't any reason for developers to call them directly and the small
documentation that was there wasn't especially helpful.

Closes #225