Encode notice message iodata before json serialize
Notice messages are sometimes provided as iodata, which may be an
improper list where the last cons is to a binary rather than a list.
That format isn't encodeable by Jason and causes a function clause error
in `Jason.Encode.list_loop/3`. This caused a secondary error to fire for
every error with an improper list.

This fixes the situation by converting iodata to binary before encoding
and sending the notice.

Closes #356