Use queue for batches, drop batch for throttling
This replaces the array with a queue so we can get FIFO behavior when
pulling batches. This is to ensure that we attempt to send the oldest
batch first.

With this, I'm also using the same batch dropping logic for throttled
batches. This way, we drop the oldest batch after 3 failures (roughly 3
minutes). When the user upgrades or the throttle is reset, the events we
send right after will be closer to the current time.