Replace timer_ref with GenServer timeouts
Based on @sorentwo's suggestion, I refactored the timer logic to lean on
GenServer timeouts, and it made the code much cleaner.

Now the GenServer callback return values contain the timeout, which we
calculate for each response. If we are throttling, we default to the 60s
throttle wait time, otherwise we use the configured timeout.

A cast or call should reset the timeout with the 3rd position of the
callback return, so each invocation essentially resets timeout.