Stop using strings to register telemetry_events
My decision to use dotified strings in the configuration to map to
instrumented telemetry events kept rubbing the wrong way. It was at the
point where I was preloading atoms to make sure the the terms were
avaliable so `to_existing_atom` would not error that the pain was
enough to reevalute things.

The config now uses lists of atoms for instrumented events. I am also
using these lists in all the places where callbacks were using the
string event name.