Add support for multiple HTTP client libraries (#579)
* Support multiple HTTP clients

Allow library consumers to choose an http client library.

Code shamelessly cribbed from [assent](https://github.com/pow-auth/assent).

* Update dummy app to include req

Since we are setting req as the default library, include it in our test
application.

* Remove httpc HTTPAdapter

In order to speed up this iteration, defer implementation of an
`:httpc` based adapter until a later time.

* Convert reduce/3 to map_join/3

Co-authored-by: Parker Selbert <parker@sorentwo.com>

* Remove redundant ifnil fallback

Co-authored-by: Parker Selbert <parker@sorentwo.com>

* Move default empty list to params

Co-authored-by: Parker Selbert <parker@sorentwo.com>

* Remove redundant downcase for headers

Co-authored-by: Parker Selbert <parker@sorentwo.com>

* Remove finch adapter

Since `req` is built on top of `finch` it is unlikely that most
consumers would want to use `finch` directly. To reduce maintenance
surface, remove the finch adapter.

* fixup: remove unused terminate callback

* Move response decode to adapter

Since req handles response body decoding for us, move the
hackney-specific response handling code into the adapter module.

* Combine 2 cases into 1 with

* Remove low-value req adapter test

This happy-path test exercises code that lives outside of this library,
so its value is low (changes in this repo are unlikely to cause it to
fail).

* Remove SSL tests

* Determine initial adapter at runtime

I noticed a few things around compiling and dependencies while testing
this:

* The compiler would complain that :hackney did not exist if I did not
  have it installed
* I would not get an error about missing HTTP library until I sent a
  request

Changes to fix:

* Refactored so the modules can load at compile time without warnings by
  using `apply` instead of bare module or atom references.
* Moved compile time default client check into a memoized function call.
* Moved the adapter check into the `Honeybadger.start/2` callback so the
  error will throw during app startup

This also ended up dealing with an issue where installing the library,
after compiling Honeybadger would not cause Honeybadger to recompile
again, which at that point it would keep erroring unless you explicitly
recompile again.

* Remove user_agent_header in adapter

The Client already adds this to the header, which was causing duplicate
headers to be transmitted.

* Req errors with nil opts

* Update readme with blurb

---------

Co-authored-by: Parker Selbert <parker@sorentwo.com>
Co-authored-by: Kevin Webster <kevin@kevinwebster.me>