Fix nested notifications issue
Since we are now using the same adapter that we are instrumenting on,
when we send an event to insights we get a telemetry event about
sending that event, and then send another event! Recursion!

This does not instrument a request event if the user-agent comes from
this library.

I also had to remove my Finch stub, since the :req dependency adds
finch.