Fix timeout issues
I did not test the timeout functionality very well in the first pass.
It was resetting the timeout on each successive event push, which is not
what I wanted.

This change keeps track of the start of the timeout and uses that to
update the GenServer with the _remaining_ timeout. Now it behaves like a
true batch limit OR timeout limit.