Insights event batching (#578)
* Basic config and client worker integration

* EventsWorker

* Fix issue where throttled batches are dropped

* Docs, spec stuff & logging dropped events

* Use queue for batches, drop batch for throttling

This replaces the array with a queue so we can get FIFO behavior when
pulling batches. This is to ensure that we attempt to send the oldest
batch first.

With this, I'm also using the same batch dropping logic for throttled
batches. This way, we drop the oldest batch after 3 failures (roughly 3
minutes). When the user upgrades or the throttle is reset, the events we
send right after will be closer to the current time.

* Apply suggestions from code review

Co-authored-by: Parker Selbert <parker@sorentwo.com>

* Move more init into the worker

* Put the old config stuff back

* Replace timer_ref with GenServer timeouts

Based on @sorentwo's suggestion, I refactored the timer logic to lean on
GenServer timeouts, and it made the code much cleaner.

Now the GenServer callback return values contain the timeout, which we
calculate for each response. If we are throttling, we default to the 60s
throttle wait time, otherwise we use the configured timeout.

A cast or call should reset the timeout with the 3rd position of the
callback return, so each invocation essentially resets timeout.

* Fix timeout issues

I did not test the timeout functionality very well in the first pass.
It was resetting the timeout on each successive event push, which is not
what I wanted.

This change keeps track of the start of the timeout and uses that to
update the GenServer with the _remaining_ timeout. Now it behaves like a
true batch limit OR timeout limit.

* Set throttle_wait from mix config

Without the default, throttle_wait was set to nil in the state.

* Remove error responses in HTTPAdapter

The client is expecting {:ok, %HTTPResponse} for even "failing"
responses (400-599). Returning error responses here was keeping the
EventWorker from reponding properly for 429 responses.

---------

Co-authored-by: Parker Selbert <parker@sorentwo.com>