Add Request ID to Insights events and notices (#597)
* Basic working Insights instrumentation

* Fill in more insights plugins & tweaks

* Put old filter_map back & remove comment

* Perhaps we should actually send the event

* Fix test & pass raw metadata to EventFilter

* Plug is doing the same thing as this.

* Also remove it from the list

* Update lib/honeybadger/insights/ecto.ex

Co-authored-by: Benjamin Curtis <ben@bencurtis.com>

* Shuffled used module specific functions around

* Add Ecto obfuscation

* Apply suggestions from code review

Co-authored-by: Parker Selbert <parker@sorentwo.com>

* Add missing end from PR feedback merge

* process_measurements/1 refactor

Thanks @sorentwo

* Add tests and class docs

This also includes the `full_url` options for tesla and finch libraries

* Update README doc

This also won't report an event if the `filter/3` returns nil.

* Load support files correctly

* Update comment for Finch module

Also ignore insights events for Tesla requests that use the Finch
adapter, as we will send 2 events.

* Fix test

* Use async

* Apply suggestions from code review

Co-authored-by: Parker Selbert <parker@sorentwo.com>

* Use `String.to_existing_atom/1`

* Load the atoms earlier

* Stop using strings to register telemetry_events

My decision to use dotified strings in the configuration to map to
instrumented telemetry events kept rubbing the wrong way. It was at the
point where I was preloading atoms to make sure the the terms were
avaliable so `to_existing_atom` would not error that the pain was
enough to reevalute things.

The config now uses lists of atoms for instrumented events. I am also
using these lists in all the places where callbacks were using the
string event name.

* Insert request_id into event payload

We are storing the request_id in the process dictionary

* Override handle_telemetry/4 & module disable

* Allows disabling a module with `config.module_name.disabled = true`
* The base module now delegates handle_telemetry to an impl function, so
  "child" modules can do the same.

* LiveView request_id for events

This stores the request_id in the process tree on either of the
:handle_event, :start events. It also clears the request_id on the stop
events, as "events" run through the same LiveView process.

* Add helper for Oban & extract request_id from job

* Fix this

* Set request_id and send start event if we see it

* Add request_id to error notice

* Inject start event in Oban for request_id

* Fix nested notifications issue

Since we are now using the same adapter that we are instrumenting on,
when we send an event to insights we get a telemetry event about
sending that event, and then send another event! Recursion!

This does not instrument a request event if the user-agent comes from
this library.

I also had to remove my Finch stub, since the :req dependency adds
finch.

* Whoops

* Remove compile time matching

* Apply suggestions from code review

Co-authored-by: Parker Selbert <parker@sorentwo.com>

* Rename uuid & fix

* Honeybadger.RequestId module & usage

* Tweaks & ecto stacktrace inclusion

* More tweaks

* Added events_* options to readme

* Add docstrings

* Apply suggestions from code review

Co-authored-by: Parker Selbert <parker@sorentwo.com>

* More PR suggestions

* Space.

---------

Co-authored-by: Benjamin Curtis <ben@bencurtis.com>
Co-authored-by: Parker Selbert <parker@sorentwo.com>