Revised event logging (#609)
* Allow Hackney case-insensitve header checking

The body was not parsing correctly for Hackney responses because it was
expecting the content type header to be lowercase.. It's not. Now we are
checking for lowercase headers and values.

* Change event logging

We just print debug counts. If there are any non 200 responses for
events, then we print them to the warn log.

* Better handling while logging errant events